Fix column users_id data type

+++ investments/database/migrations/2021_09_29_211010_investiments.php
@@ -15,7 +15,7 @@ class Investiments extends Migration
     {
         Schema::create('investments', function (Blueprint $table) {
             $table->id();
-            $table->integer('users_id');
+            $table->bigInteger('users_id');
             $table->integer('amount');
             $table->date('date');
             $table->timestamps();
